Fixing the file structure according to issue #58
