+++ dLux/detectors.py
