+++ dLux/dev/__init__.py
@@ -1,0 +1,1 @@
+name = "dev"
