+++ dLux/utils/__init__.py
@@ -1,0 +1,12 @@
+name = "utils"
+
+from .plotting import plot_batch
+from .zernike import zernike_basis
+from .hexike import hexike_basis
+from .helpers import (unif_rand, norm_rand, opd2phase, phase2opd,
+    get_ppf, rad2arcsec, arcsec2rad, cart2polar, polar2cart,
+    scale_mask, get_GE, get_RGE, get_RWGE, get_Rvec, get_Rmask,
+    nyquist_pix_airy, nyquist_pix)
+from .bayes import (poiss_logpri, chi2_logpri, calc_cov, calc_ent)
+from .coordinates import (get_pixel_vector, get_pixel_positions,
+    get_radial_positions)
