--- src/dLux/__init__.py
@@ -1,6 +1,0 @@
-name = "dLux"
-
-import dLux.optics
-import dLux.statistics
-import dLux.graphics
-import dLux.utilities
