--- src/dLux/graphics/__init__.py
@@ -1,3 +1,0 @@
-name = "graphics"
-
-from .plotting import *
