--- src/dLux/statistics/__init__.py
@@ -1,5 +1,0 @@
-name = "statistics"
-
-from .zernike import *
-from .hexike import *
-from .bayes import *
