--- src/dLux/utilities/__init__.py
@@ -1,5 +1,0 @@
-name = "utilities"
-
-from .helpers import *
-from .image import *
-from .models import *
